third_party/blink/web_tests/fast/html/imports/rel-style-to-import-expected.txt

CONSOLE WARNING: line 18: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: CSS cannot be loaded from `file:` URLs unless they end in a `.css` file extension.
Updating attribute value shouldn't result import loading.
Before update:
PASS target.sheet is non-null.
PASS target.import is null
After update:
PASS target.sheet is non-null.
PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
